Extract standard type ext. not depending on Gradle
